feat: move ConnectionManager from slipp into nextstep-jdbc
