+++ nextstep-jdbc/src/main/java/nextstep/jdbc/ConnectionManager.java
@@ -1,4 +1,4 @@
-package slipp.support.db;
+package nextstep.jdbc;
 
 import org.apache.commons.dbcp2.BasicDataSource;
 
@@ -7,17 +7,24 @@ import java.sql.Connection;
 import java.sql.SQLException;
 
 public class ConnectionManager {
-    private static final String DB_DRIVER = "org.h2.Driver";
-    private static final String DB_URL = "jdbc:h2:mem:jwp-framework";
-    private static final String DB_USERNAME = "sa";
-    private static final String DB_PW = "";
+    private static String DRIVER;
+    private static String URL;
+    private static String USERNAME;
+    private static String PASSWORD;
+
+    public static void initialize(String _driver, String url, String userName, String password) {
+        DRIVER = _driver;
+        URL = url;
+        USERNAME = userName;
+        PASSWORD = password;
+    }
 
     public static DataSource getDataSource() {
         BasicDataSource ds = new BasicDataSource();
-        ds.setDriverClassName(DB_DRIVER);
-        ds.setUrl(DB_URL);
-        ds.setUsername(DB_USERNAME);
-        ds.setPassword(DB_PW);
+        ds.setDriverClassName(DRIVER);
+        ds.setUrl(URL);
+        ds.setUsername(USERNAME);
+        ds.setPassword(PASSWORD);
         return ds;
     }
